pyproject.toml
[project]
name = "energy-analysis"
version = "0.1.0"
description = "A data science project for analyzing weather and energy usage"
authors = [{ name = "Energy Analysis Group" }]
requires-python = ">=3.10"

dependencies = [
    "pandas>=2.2.3",
    "numpy>=1.26.4",
    "scikit-learn>=1.6.1",
    "matplotlib>=3.8.4",
    "seaborn>=0.13.2",
    "plotly>=5.22.0",
    "streamlit>=1.30.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "tqdm>=4.67.1",
    "loguru>=0.7.3",
    "statsmodels>=0.14.1" 
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
    "deptry>=0.23.0",
    "ipykernel>=6.29.5"
]

[tool.ruff]
line-length = 100
fix = true
target-version = "py310"

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.ruff.lint]
extend-select = ["E", "W", "F", "B", "I", "UP", "SIM", "C4", "N", "ARG"]
ignore = ["S101", "ANN401", "ERA001"]

[tool.ruff.lint.isort]
known-first-party = ["energy_analysis"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]

[tool.mypy]
files = ["src"]
explicit_package_bases = true
mypy_path = ["src"]
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
python_version = "3.10"

[[tool.mypy.overrides]]
module = ["requests.*", "streamlit.*"]
ignore_missing_imports = true

[tool.deptry]
exclude = [".venv"]
known_first_party = ["energy_analysis"]

[tool.deptry.per_rule_ignores]
DEP004 = ["pytest"]
